refactor: Separated interface for queries responses.
